pytest: Separate task tests
